refactor(room-details): 컴포넌트 분리 (#23)
* refactor(room-details): indicator 리팩토링

* refactor: 컴포넌트 분리
