+++ src/features/checklist/components/index.tsx
@@ -1,2 +1,13 @@
 import ChecklistView from "./ChecklistView";
-export { ChecklistView };
+import QuestionItem from "./QuestionItem";
+import SingleChoiceView from "./SingleChoiceView";
+import MultipleChoiceView from "./MultipleChoiceView";
+import QuestionPanel from "./QuestionPanel";
+
+export {
+  ChecklistView,
+  QuestionItem,
+  MultipleChoiceView,
+  SingleChoiceView,
+  QuestionPanel,
+};
