+++ src/features/rooms/components/index.ts
@@ -3,6 +3,7 @@ import DropDown from "./DropDown";
 import RoomListView from "./RoomListView";
 import DetailCard, { SummaryTable } from "./DetailCard";
 import { DeleteModal } from "./DeleteModal";
+import RoomCardIndicator from "./RoomCardIndicator";
 
 export {
   RoomCard,
@@ -11,5 +12,6 @@ export {
   RoomListView,
   DetailCard,
   SummaryTable,
+  RoomCardIndicator,
   DeleteModal,
 };
